Unify detail-pane focus/navigation key model between Issues and PR modes (Fixes #150)
The Issues and PR detail panes used inconsistent keyboard models for the same
conceptual operations: in Issues detail Tab moved focus within the detail
(subfocus), while in PR detail Tab moved focus out of the detail (next pane).
PR detail subfocus was only reachable via j/k, making it undiscoverable.

Unify on the maintainer-preferred model: arrows own spatial navigation
(panels + items); Tab owns within-pane detail-subfocus cycling; j/k is
retained as a subfocus alias in both detail panes for vim users.

Key changes:
- PR detail (handle_pr_detail_key): Tab/BackTab now cycle detail subfocus
  (PrDetailSubfocusNext/Prev), intercepting before the pane-cycle fallback, so
  PR detail matches Issues. j/k kept as aliases. Left/Right now symmetrically
  cycle panes (Right was previously unbound).
- Issues detail (resolve_issue_detail_key_event): j/k added as subfocus
  aliases (Tab/BackTab already did subfocus). Right arrow added for symmetric
  pane-focus.
- List/repo panes keep Tab as a pane-cycle alias (no internal subfocus).
- Keybind bar (keybind_hints_for): both modes now describe the real model
  (^/v items, </> panes, Tab detail focus (j/k)) plus PR-specific actions.
- Help modal: rewrote content to document the unified model and the
  review-thread workflow (Tab/j/k to a thread, then R resolve / r reply);
  restored previously-valid bindings (Space, v, 1/2/3). Wired the
  pre-existing-but-dead help scroll plumbing so the modal scrolls via
  ScrollableText, sized to never overflow short terminals.

Tests: updated existing key-routing tests for the unified PR-detail Tab=subfocus
behavior; added new behavioral tests for j/k aliases, Tab/BackTab subfocus, and
symmetric pane arrows; added pure sizing tests for the help viewport helper.
